Video functionality
You have helped me to build a nice audio/transcript synchroniser. Have a read of README.md to understand what has been built.

Now, I want to build a .MP4/transcript synchroniser. Exactly the same functionality, but instead of just playing a .MP3, I want there be a video player. Here’s what I want you to do:

1. Update README.md to reflect the functionality I am going to talk about below. Leave the .MP3 functionality
2. Update the first page to allow for MP4 files to be uploaded. Use the same button at the top, but just change the label text and the current file format restriction. Do not change anything about the 2nd page.
3. If a user uploads a MP3 file, take them to the current 2nd page interface.
4. If a user uploads a MP4 files take them to a new 2nd page interface. The right hand side, from top to bottom should be the transcript player. The words should highlight as the speaker speaks, and if a user clicks further down the page, the player (more on that later) should jump to that new position - exactly like the current MP3/transcript functionality. The left hand side should be the video player. The video player should display the video, and if a user jumps to a new position, the video should update. The video should be as far left and right as the box allows, and centred. Finally, the toolbar that is currently at the top of the page in the MP3 product, should be at the bottom left, under the video. All functions that are in the toolbar should also be in this toolbar and work the same.

DO NOT change the functionality of the MP3 part of the product. You are only adding a new part of the product, and a new part of the logic tree on the 1st page. It is especially important that the MP3 part of the product still works.

Think step by step, let me know if you have questions.